# Conflicts:
#	IrrigationAdvisor/IrrigationAdvisor.csproj
#	IrrigationAdvisor/Models/Agriculture/Specie.cs
#	IrrigationAdvisorConsole/Program.cs

+++ IrrigationAdvisor.WebApi/Models/FarmViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class FarmViewModel
     {
-        public int FarmId { get; set; }
+        public long FarmId { get; set; }
         public string Description { get; set; }
     }
 }
